Split PanZoom mesh in Pan and Zoom meshes
